+++ tests/bootstrap.php
@@ -1,2 +1,4 @@
 <?php
-	require dirname(__FILE__)."/../autoload.php";
+	ini_set("display_errors", 1);
+	error_reporting(E_ALL);
+	require __DIR__."/../autoload.php";